requirements.txt
pypdf2~=3.0.1
python-docx~=1.2.0
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
attrs~=25.1.0
boto3~=1.38.46
boto3-stubs~=1.40.20
botocore~=1.38.46
botocore-stubs~=1.38.46
certifi~=2025.1.31
charset-normalizer~=3.4.3
colorama~=0.4.6
et_xmlfile~=2.0.0
execnet~=2.1.1
gherkin-official~=29.0.0
greenlet~=3.1.1
idna~=3.10
iniconfig~=2.0.0
Jinja2~=3.1.6
jmespath~=1.0.1
Mako~=1.3.10
MarkupSafe~=3.0.2
numpy~=2.3.2
openpyxl~=3.1.5
packaging~=24.2
pandas~=2.3.2
parse~=1.20.2
parse_type~=0.6.6
playwright~=1.49.1
pluggy~=1.5.0
pyee~=12.0.0
Pygments~=2.19.2
pyodbc~=5.2.0
pytest~=8.3.5
pytest-base-url~=2.1.0
pytest-bdd~=8.1.0
pytest-html~=4.1.1
pytest-metadata~=3.1.1
pytest-order~=1.3.0
pytest-playwright~=0.7.0
pytest-rerunfailures~=15.1
pytest-timeout~=2.4.0
pytest-xdist~=3.6.1
python-dateutil~=2.9.0.post0
python-dotenv~=1.1.1
python-slugify~=8.0.4
pytz~=2025.2
PyYAML~=6.0.2
requests~=2.32.5
s3transfer~=0.13.1
six~=1.17.0
SQLAlchemy~=2.0.43
text-unidecode~=1.3
types-awscrt~=0.27.6
types-s3transfer~=0.13.0
typing_extensions~=4.12.2
tzdata~=2025.2
urllib3~=2.5.0
